+++ dsfhub/resource_log_aggregator.go
@@ -582,8 +582,8 @@ func resourceLogAggregatorRead(d *schema.ResourceData, m interface{}) error {
 	d.Set("server_type", logAggregatorReadResponse.Data.ServerType)
 	if logAggregatorReadResponse.Data.AssetData.ServerPort != nil {
 		var serverPort string
-		if serverPortNum, ok := logAggregatorReadResponse.Data.AssetData.ServerPort.(int); ok {
-			serverPort = fmt.Sprintf("%d", int(serverPortNum))
+		if serverPortNum, ok := logAggregatorReadResponse.Data.AssetData.ServerPort.(float64); ok {
+			serverPort = fmt.Sprintf("%d", float64(serverPortNum))
 		} else {
 			serverPort = logAggregatorReadResponse.Data.AssetData.ServerPort.(string)
 		}
