+++ dsfhub/resource_secret_manager.go
@@ -551,8 +551,8 @@ func resourceSecretManagerRead(d *schema.ResourceData, m interface{}) error {
 	d.Set("server_ip", secretManagerReadResponse.Data.AssetData.ServerIP)
 	if secretManagerReadResponse.Data.AssetData.ServerPort != nil {
 		var serverPort string
-		if serverPortNum, ok := secretManagerReadResponse.Data.AssetData.ServerPort.(int); ok {
-			serverPort = fmt.Sprintf("%d", int(serverPortNum))
+		if serverPortNum, ok := secretManagerReadResponse.Data.AssetData.ServerPort.(float64); ok {
+			serverPort = fmt.Sprintf("%d", float64(serverPortNum))
 		} else {
 			serverPort = secretManagerReadResponse.Data.AssetData.ServerPort.(string)
 		}
